% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{attributes}
\alias{attributes}
\title{Combine attributes}
\usage{
attributes(...)
}
\arguments{
\item{...}{Strings.}
}
\value{
Combined string with all the attributes.
}
\description{
Combine attributes from a vector of strings.
}
\examples{
dabr::attributes("A", "B", "C")
dabr::attributes(c("A", "B", "C"))
dabr::attributes(c("A", "B", "C"), "D", "E", "F")
}
