% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\docType{package}
\name{dabr-package}
\alias{dabr}
\alias{dabr-package}
\title{dabr: Database Management with R}
\description{
Provides functions to manage databases: select, update, insert, and delete records, list tables, backup tables as CSV files, and import CSV files as tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/special-uor/dabr/}
  \item \url{https://special-uor.github.io/dabr/}
  \item \url{https://research.reading.ac.uk/palaeoclimate/}
  \item Report bugs at \url{https://github.com/special-uor/dabr/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@outlook.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Other contributors:
\itemize{
  \item SPECIAL Research Group @ University of Reading [copyright holder]
}

}
\keyword{internal}
