% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{get_attr}
\alias{get_attr}
\alias{get_attr.MariaDBConnection}
\title{Get attributes of a table}
\usage{
get_attr(conn, ...)

\method{get_attr}{MariaDBConnection}(conn, name, ...)
}
\arguments{
\item{conn}{DB connection object.}

\item{...}{Optional parameters.}

\item{name}{Table name.}
}
\value{
List of attributes for table \code{name}.
}
\description{
Get attributes of a table
}
