% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{is.connected}
\alias{is.connected}
\alias{is.connected.MariaDBConnection}
\title{Verify connection}
\usage{
is.connected(conn, ...)

\method{is.connected}{MariaDBConnection}(conn, ...)
}
\arguments{
\item{conn}{DB connection object.}

\item{...}{Optional parameters.}
}
\value{
Connection status.
}
\description{
Verify if connection object is still valid, is connected to the database
server.
}
