% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{open_conn_mysql}
\alias{open_conn_mysql}
\title{Connect to database}
\usage{
open_conn_mysql(dbname, user, password = NULL, host = "localhost", port = 3306)
}
\arguments{
\item{dbname}{database name}

\item{user}{username of database owner}

\item{password}{password \link{default: \code{NULL}}}

\item{host}{database host, it can be local (default) or remote}

\item{port}{database port}
}
\value{
\code{MariaDBConnection} connection object
}
\description{
Uses \code{RMariaDB} to open a connection to a MySQL database
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
}

}
\seealso{
Other DB functions: 
\code{\link{close_conn}()},
\code{\link{select_all}()},
\code{\link{select_query_mysql}()}
}
\concept{DB functions}
