% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{quote}
\alias{quote}
\title{Add single quotes}
\usage{
quote(str)
}
\arguments{
\item{str}{String.}
}
\value{
String surrounded by single quotes
}
\description{
Add single quotes to string.
}
\examples{
dabr::quote("A")
dabr::quote("l'A")
}
