% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{select}
\alias{select}
\alias{select.MariaDBConnection}
\alias{update}
\title{Execute \code{SELECT} query}
\usage{
select(conn, ...)

\method{select}{MariaDBConnection}(conn, query, quiet = FALSE, ...)

update(conn, ...)
}
\arguments{
\item{conn}{\code{MariaDBConnection} connection object.}

\item{...}{Optional parameters.}

\item{query}{\code{SELECT} query.}

\item{quiet}{Boolean flag to hide status messages.}
}
\value{
Data frame containing the selected records.
}
\description{
Execute \code{SELECT} query

Execute \code{UPDATE} query
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
out <- select(conn, "SELECT variable, value FROM sys_config")
close_conn(conn)
}
}
\seealso{
Other DB functions: 
\code{\link{close_conn}()},
\code{\link{open_conn_mysql}()},
\code{\link{select_all}()}

Other DB functions: 
\code{\link{close_conn}()},
\code{\link{open_conn_mysql}()},
\code{\link{select_all}()}
}
\concept{DB functions}
