% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{select_query_mysql}
\alias{select_query_mysql}
\title{Execute select query}
\usage{
select_query_mysql(conn, query, quiet = FALSE)
}
\arguments{
\item{conn}{\code{MariaDBConnection} connection object}

\item{query}{select query}

\item{quiet}{boolean flag to hide status messages}
}
\value{
data frame containing the selected records
}
\description{
Retrieves records that meet the criteria declared in the query
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
out <- select_query_mysql(conn, "SELECT variable, value FROM sys_config")
close_conn(conn)
}
}
\seealso{
Other DB functions: 
\code{\link{close_conn}()},
\code{\link{open_conn_mysql}()},
\code{\link{select_all}()}
}
\concept{DB functions}
