% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{close_conn}
\alias{close_conn}
\alias{close_conn.MariaDBConnection}
\alias{close_conn.default}
\title{Close connection to database}
\usage{
close_conn(conn, ...)

\method{close_conn}{MariaDBConnection}(conn, ...)

\method{close_conn}{default}(conn, ...)
}
\arguments{
\item{conn}{Connection object.}

\item{...}{Optional parameters.}
}
\description{
Close connection to database
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
close_conn(conn)
}
}
\seealso{
Other DB functions: 
\code{\link{insert}()},
\code{\link{list_tables}()},
\code{\link{open_conn_mysql}()},
\code{\link{select_all}()},
\code{\link{select}()},
\code{\link{update}()}
}
\concept{DB functions}
