% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{list_tables}
\alias{list_tables}
\alias{list_tables.MariaDBConnection}
\title{List tables}
\usage{
list_tables(conn, ...)

\method{list_tables}{MariaDBConnection}(conn, quiet = FALSE, attr = TRUE, ...)
}
\arguments{
\item{conn}{DB connection object.}

\item{...}{Optional parameters.}

\item{quiet}{Boolean flag to hide status messages.}

\item{attr}{Boolean flag to list the attributes of each table.}
}
\value{
If \code{quiet = TRUE} returns a list with the tables' names. If
\code{attr = TRUE} includes each attribute of the tables.
}
\description{
List tables
}
\seealso{
Other DB functions: 
\code{\link{close_conn}()},
\code{\link{delete}()},
\code{\link{insert}()},
\code{\link{open_conn_mysql}()},
\code{\link{select_all}()},
\code{\link{select}()},
\code{\link{update}()}
}
\concept{DB functions}
