% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{open_conn_mysql}
\alias{open_conn_mysql}
\title{Connect to database}
\usage{
open_conn_mysql(
  dbname,
  user = "root",
  password = NULL,
  host = "127.0.0.1",
  port = 3306
)
}
\arguments{
\item{dbname}{Database/Schema name.}

\item{user}{Username of database owner.}

\item{password}{Password (default: \code{NULL}).}

\item{host}{Database host, it can be local (default) or remote.}

\item{port}{Database port.}
}
\value{
\code{MariaDBConnection} connection object.
}
\description{
Uses \code{RMariaDB} to open a connection to a MySQL database.
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys")
}

}
\seealso{
Other DB functions: 
\code{\link{close_conn}()},
\code{\link{delete}()},
\code{\link{insert}()},
\code{\link{list_tables}()},
\code{\link{select_all}()},
\code{\link{select}()},
\code{\link{update}()}
}
\concept{DB functions}
