% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabr.R
\name{update}
\alias{update}
\alias{update.MariaDBConnection}
\title{Execute \code{UPDATE} query}
\usage{
update(conn, ...)

\method{update}{MariaDBConnection}(conn, query, quiet = FALSE, ...)
}
\arguments{
\item{conn}{\code{MariaDBConnection} connection object.}

\item{...}{Optional parameters.}

\item{query}{\code{UPDATE} query.}

\item{quiet}{Boolean flag to hide status messages.}
}
\description{
Execute \code{UPDATE} query
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
out <- update(conn, "UPDATE sys_config SET value = 1")
close_conn(conn)
}
}
\seealso{
Other DB functions: 
\code{\link{close_conn}()},
\code{\link{insert}()},
\code{\link{list_tables}()},
\code{\link{open_conn_mysql}()},
\code{\link{select_all}()},
\code{\link{select}()}
}
\concept{DB functions}
